requirements.txt
aiofiles
aioshutil
anytree
apscheduler
aria2p
asyncio
cloudscraper
dnspython
feedparser
flask
gevent
httpx
lxml
motor
natsort
pillow
psutil
pymongo
pyrogram
python-dotenv
python-magic
requests
tenacity
tgcrypto
urllib3
uvloop
xattr
yt-dlp
